%
\docType{data}
\name{wq}
\alias{wq}
\title{Water Quality Measurements}
\format{A data frame with  1033 observations and 15 variables
\describe{
\item{Date}{Date}
\item{Depth}{meter}
\item{EC}{electro conductivity}
\item{Station}{station}
\item{Temp}{water temperature}}}
\source{
digitalizado
}
\usage{
wq
}
\description{
water quality measurements.
}
\keyword{datasets}
